+++ include/platform/posix/net_impl.h
@@ -174,5 +174,6 @@ static inline int someip_socket_errno() { return errno; }
 #define SOMEIP_EINPROGRESS EINPROGRESS
 #define SOMEIP_EBADF       EBADF
 #define SOMEIP_EINTR       EINTR
+#define SOMEIP_ETIMEDOUT   ETIMEDOUT
 
 #endif // SOMEIP_PLATFORM_POSIX_NET_IMPL_H